refactor: remove useless files, remove useless dependencies
