+++ package.json
@@ -15,8 +15,6 @@
     "@hookform/resolvers": "3.3.2",
     "@reduxjs/toolkit": "1.9.7",
     "classnames": "2.3.2",
-    "i": "0.3.7",
-    "npm": "10.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "7.48.2",
