--- src/pages/ControlledForm/ControlledForm.tsx
@@ -1,5 +1,0 @@
-const ControlledForm = () => {
-  return <div>Controlled form</div>;
-};
-
-export default ControlledForm;
