--- src/pages/UncontrolledForm/UncontrolledForm.tsx
@@ -1,5 +1,0 @@
-const UncontrolledForm = () => {
-  return <div>Uncontrolled Form</div>;
-};
-
-export default UncontrolledForm;
